Rename polyfill files to match Google TS style guide.
