+++ packages/formdata-event/ts_src/environment/window.ts
